change root command from config to client
